+++ datatypes/src/commonMain/kotlin/at/asitplus/crypto/datatypes/asn1/Asn1Encoder.kt
@@ -10,6 +10,7 @@ import at.asitplus.crypto.datatypes.asn1.BERTags.NULL
 import at.asitplus.crypto.datatypes.asn1.BERTags.OBJECT_IDENTIFIER
 import at.asitplus.crypto.datatypes.asn1.BERTags.OCTET_STRING
 import at.asitplus.crypto.datatypes.asn1.BERTags.UTC_TIME
+import at.asitplus.crypto.datatypes.io.BitSet
 import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
@@ -97,9 +98,16 @@ class Asn1TreeBuilder {
     fun octetString(child: Asn1Element) = apply { octetString(block = { child.derEncoded }) }
 
     /**
-     * Adds the passed bytes as BIT STRING [Asn1Primitive] to this ASN.1 structure
+     * Adds the passed bytes as BIT STRING [Asn1Primitive] to this ASN.1 structure.
+     * **Right-aligned, i.e. prepended with `0x00`**
      */
     fun bitString(block: () -> ByteArray) = apply { elements += block().encodeToTlvBitString() }
+
+    /**
+     * Transforms the passed BitSet as BIT STRING [Asn1Primitive] to this ASN.1 structure.
+     * **Left-Aligned and right-padded (see [Asn1BitString])**
+     */
+    fun bitString(bitSet: BitSet) = apply { elements += Asn1BitString(bitSet).encodeToTlv() }
 
     /**
      * Adds the passed [Asn1Element] as BIT STRING [Asn1Primitive] to this ASN.1 structure
@@ -320,12 +328,7 @@ fun ByteArray.encodeToTlvOctetString() = Asn1Primitive(OCTET_STRING, this)
 /**
  * Produces a BIT STRING as [Asn1Primitive]
  */
-fun ByteArray.encodeToTlvBitString() = Asn1Primitive(BIT_STRING, encodeToBitString())
-
-/**
- * Prepends 0x00 to this ByteArray for encoding it into a BIT STRING. Useful for implicit tagging
- */
-fun ByteArray.encodeToBitString() = byteArrayOf(0x00) + this
+private fun ByteArray.encodeToTlvBitString() = Asn1Primitive(BIT_STRING,  byteArrayOf(0x00, *this))
 
 private fun String.encodeTolvOid() = Asn1Primitive(OBJECT_IDENTIFIER, decodeToByteArray(Base16()))